+++ src/Data/Commits/CommitAuthorData.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace JordanPartridge\GithubClient\Data;
+namespace JordanPartridge\GithubClient\Data\Commits;
 
 use Carbon\Carbon;
 use Spatie\LaravelData\Data;
